Move relationship helper class
The class is moved so that it does not get imported with all models,
since it has to be loaded after all other modules are available. That is
so we can have helper classes that handle special relationships in other
modules as well, such as workflow and other workflow objects.
